requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.30.0


# Bleeding edge Django
django==1.11.7

# Configuration
django-environ==0.4.4
whitenoise==3.3.1


# Models
django-model-utils==3.0.0
django-extensions==1.9.8

# Images
Pillow==4.3.0

# Password storage
argon2-cffi==16.3.0

# Python-PostgreSQL Database Adapter
psycopg2==2.7.4

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2017.3

# Redis support
django-redis==4.8.0
redis>=2.10.5


celery==4.1.0

# Your custom requirements go here

slacker==0.9.60
graphene-django==2.0.0
django-cors-headers==2.1.0
djangorestframework-camel-case==0.2.0
wagtail==1.13.1
beautifulsoup4==4.6.0
requests==2.18.4
fuzzywuzzy[speedup]
mailchimp3==2.0.18
arrow==0.12.1
djangorestframework-jwt==1.11.0
boto3==1.5.28
PyJWT==1.5.3
django-filter
